:recycle: Moved the ComparisonSectionSelector component from the Main component to the ComparisonPanel component. :lipstick: Adapted the styling to the refactor.
