+++ src/assets/styles/fonts-panel.module.css
@@ -1,6 +1,7 @@
 #fontsPanel {
 	grid-column: 1/7;
 	height: var(--fonts-panel-height);
+	z-index: 10;
 }
 
 .fontSelectorContainer {
